# saas-forge-contracts/compatibility-baselines/v1/v0.1.0/openapi/v1.yaml
openapi: 3.1.1
info:
  title: saas-forge Public API
  version: 1.0.0
  description: |
    The versioned public REST contract for saas-forge. This document contains only
    delivery-stage-two and delivery-stage-three resources. Permission, Feature,
    and Quota runtime operations are intentionally absent until their domain
    contracts are reviewed.
servers:
  - url: https://api.example.invalid
    description: Deployment-specific API Gateway origin.
...
paths:
...
  /api/v1/platform/quota-definitions:
    post:
      tags: [Platform entitlement bootstrap]
      operationId: createQuotaDefinition
      x-saas.forge-service: entitlement-service
      summary: Create the max_users Quota Definition in DRAFT state
      security: [{ UserBearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateQuotaDefinitionRequest' }
      responses:
        '201':
          description: The Quota Definition was created in DRAFT state.
          headers:
            Location: { $ref: './common.yaml#/components/headers/Location' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QuotaDefinition' }
        '400': { $ref: './common.yaml#/components/responses/ValidationProblem' }
        '401': { $ref: './common.yaml#/components/responses/Problem' }
        '403': { $ref: './common.yaml#/components/responses/Problem' }
        '409': { $ref: './common.yaml#/components/responses/Problem' }
  /api/v1/platform/quota-definitions/{quotaDefinitionId}/activations:
    post:
      tags: [Platform entitlement bootstrap]
      operationId: activateQuotaDefinition
      x-saas.forge-service: entitlement-service
      summary: Activate a DRAFT Quota Definition
      security: [{ UserBearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/QuotaDefinitionId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: The Quota Definition is ACTIVE.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QuotaDefinition' }
        '400': { $ref: './common.yaml#/components/responses/ValidationProblem' }
        '401': { $ref: './common.yaml#/components/responses/Problem' }
        '403': { $ref: './common.yaml#/components/responses/Problem' }
        '404': { $ref: './common.yaml#/components/responses/Problem' }
        '409': { $ref: './common.yaml#/components/responses/Problem' }
  /api/v1/platform/plans:
    post:
      tags: [Platform entitlement bootstrap]
      operationId: createPlan
      x-saas.forge-service: entitlement-service
      summary: Create a DRAFT Plan with its max_users limit
      security: [{ UserBearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePlanRequest' }
      responses:
        '201':
          description: The Plan was created in DRAFT state.
          headers:
            Location: { $ref: './common.yaml#/components/headers/Location' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Plan' }
        '400': { $ref: './common.yaml#/components/responses/ValidationProblem' }
        '401': { $ref: './common.yaml#/components/responses/Problem' }
        '403': { $ref: './common.yaml#/components/responses/Problem' }
        '409': { $ref: './common.yaml#/components/responses/Problem' }
  /api/v1/platform/plans/{planId}/activations:
    post:
      tags: [Platform entitlement bootstrap]
      operationId: activatePlan
      x-saas.forge-service: entitlement-service
      summary: Activate a DRAFT Plan
      security: [{ UserBearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/PlanId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: The Plan is ACTIVE.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Plan' }
        '400': { $ref: './common.yaml#/components/responses/ValidationProblem' }
        '401': { $ref: './common.yaml#/components/responses/Problem' }
        '403': { $ref: './common.yaml#/components/responses/Problem' }
        '404': { $ref: './common.yaml#/components/responses/Problem' }
        '409': { $ref: './common.yaml#/components/responses/Problem' }
  /api/v1/platform/tenants/{tenantId}/subscriptions:
    post:
      tags: [Platform entitlement bootstrap]
      operationId: createInitialSubscription
      x-saas.forge-service: entitlement-service
      summary: Create the first active Subscription for a Tenant
      description: Requires a Platform Admin and an ACTIVE Plan. This batch supports only immediate ACTIVE Subscriptions.
      security: [{ UserBearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateInitialSubscriptionRequest' }
      responses:
        '201':
          description: The first active Subscription was created.
          headers:
            Location: { $ref: './common.yaml#/components/headers/Location' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Subscription' }
        '400': { $ref: './common.yaml#/components/responses/ValidationProblem' }
        '401': { $ref: './common.yaml#/components/responses/Problem' }
        '403': { $ref: './common.yaml#/components/responses/Problem' }
        '404': { $ref: './common.yaml#/components/responses/Problem' }
        '409': { $ref: './common.yaml#/components/responses/Problem' }
...
components:
  securitySchemes:
    UserBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: A Platform Admin user Access Token. OAuth Client tokens cannot access platform management operations.
    OAuthClientBasic:
      type: http
      scheme: basic
      description: OAuth Client client_id and client_secret.
...
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { $ref: './common.yaml#/components/schemas/UuidV7' }
...
    TenantId:
      name: tenantId
      in: path
      required: true
      schema: { $ref: './common.yaml#/components/schemas/UuidV7' }
    QuotaDefinitionId:
      name: quotaDefinitionId
      in: path
      required: true
      schema: { $ref: './common.yaml#/components/schemas/UuidV7' }
    PlanId:
      name: planId
      in: path
      required: true
      schema: { $ref: './common.yaml#/components/schemas/UuidV7' }
...
  schemas:
    Timestamp:
      type: string
      format: date-time
      pattern: '^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}[.][0-9]{3}Z$'
...
    QuotaDefinitionStatus:
      type: string
      enum: [DRAFT, ACTIVE, RETIRED]
    QuotaDefinition:
      type: object
      required: [id, code, status, createdAt, updatedAt]
      properties:
        id: { $ref: './common.yaml#/components/schemas/UuidV7' }
        code: { type: string, const: max_users }
        status: { $ref: '#/components/schemas/QuotaDefinitionStatus' }
        createdAt: { $ref: '#/components/schemas/Timestamp' }
        updatedAt: { $ref: '#/components/schemas/Timestamp' }
      additionalProperties: false
    CreateQuotaDefinitionRequest:
      type: object
      required: [code]
      properties:
        code: { type: string, const: max_users }
      additionalProperties: false
    PlanStatus:
      type: string
      enum: [DRAFT, ACTIVE, RETIRED]
    PlanQuotaLimit:
      type: object
      required: [quotaDefinitionId, limit]
      properties:
        quotaDefinitionId: { $ref: './common.yaml#/components/schemas/UuidV7' }
        limit: { type: integer, minimum: 0 }
      additionalProperties: false
    Plan:
      type: object
      required: [id, code, displayName, status, quotaLimits, createdAt, updatedAt]
      properties:
        id: { $ref: './common.yaml#/components/schemas/UuidV7' }
        code:
          type: string
          pattern: '^[a-z][a-z0-9-]{1,62}$'
        displayName: { type: string, minLength: 1, maxLength: 200 }
        status: { $ref: '#/components/schemas/PlanStatus' }
        quotaLimits:
          type: array
          minItems: 1
          maxItems: 1
          items: { $ref: '#/components/schemas/PlanQuotaLimit' }
        createdAt: { $ref: '#/components/schemas/Timestamp' }
        updatedAt: { $ref: '#/components/schemas/Timestamp' }
      additionalProperties: false
    CreatePlanRequest:
      type: object
      required: [code, displayName, quotaLimits]
      properties:
        code:
          type: string
          pattern: '^[a-z][a-z0-9-]{1,62}$'
        displayName: { type: string, minLength: 1, maxLength: 200 }
        quotaLimits:
          type: array
          minItems: 1
          maxItems: 1
          items: { $ref: '#/components/schemas/PlanQuotaLimit' }
      additionalProperties: false
    Subscription:
      type: object
      required: [id, tenantId, planId, status, endsAt, createdAt]
      properties:
        id: { $ref: './common.yaml#/components/schemas/UuidV7' }
        tenantId: { $ref: './common.yaml#/components/schemas/UuidV7' }
        planId: { $ref: './common.yaml#/components/schemas/UuidV7' }
        status: { type: string, const: ACTIVE }
        endsAt:
          oneOf:
            - $ref: '#/components/schemas/Timestamp'
            - type: 'null'
        createdAt: { $ref: '#/components/schemas/Timestamp' }
      additionalProperties: false
    CreateInitialSubscriptionRequest:
      type: object
      required: [planId]
      properties:
        planId: { $ref: './common.yaml#/components/schemas/UuidV7' }
        endsAt:
          oneOf:
            - $ref: '#/components/schemas/Timestamp'
            - type: 'null'
      additionalProperties: false
...